pyproject.toml
[project]
name = "skku-kdt-opensearch-25"
version = "0.1.0"
description = "AWS Bedrock과 OpenSearch를 활용한 나만의 검색형 AI 만들기"
authors = [
    {name = "김한성", email = "gnidoc327@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "boto3 (>=1.38.46,<2.0.0)",
    "opensearch-py (>=2.8.0,<3.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "chainlit (>=2.9.6,<3.0.0)",
    "langchain-mcp-adapters (>=0.2.1,<1.0.0)",
    "langchain-aws (>=1.2.3,<2.0.0)",
    "langgraph (>=1.0.8,<2.0.0)",
    "ddgs (>=9.10.0,<10.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
